Registry simplification.
